Add packages to db and to tab
